versao final ajuste alerta permissao pagina proibida e lista de usuarios cadastrados com data login criado e ultimo acesso

+++ gerenciador_industriawayne/views.py
@@ -13,19 +13,29 @@ from django.core.paginator import Paginator
 def home(request):
     return render(request, 'gerenciador_industriawayne/home.html')
 @login_required(login_url='/login/login')
-@permission_required('gerenciador_industriawayne.add_equipamentos', raise_exception=True)
 def cadastrar_equipamento(request):
-    form = EquipamentosForms(request.POST, request.FILES)
-    if form.is_valid():
-        form.save()
-        messages.success(request, 'Equipamento Cadastrado com sucesso!')
-    return render(request, 'gerenciador_industriawayne/cadastrar_equipamento.html', {'form': form})
+
+    if not request.user.has_perm('gerenciador_industriawayne.add_equipamentos'):
+        # Adiciona uma mensagem de erro
+        messages.error(request, 'Você não tem permissão para acessar esta página.')
+        return redirect('home')
+    else:
+        form = EquipamentosForms(request.POST, request.FILES)
+        if form.is_valid():
+            form.save()
+            messages.success(request, 'Equipamento Cadastrado com sucesso!')
+        return render(request, 'gerenciador_industriawayne/cadastrar_equipamento.html', {'form': form})
 
 @login_required(login_url='/login/login')
-@permission_required('gerenciador_industriawayne.change_equipamentos', raise_exception=True)
 def listar_equipamentos(request):
-    equipamentos = Equipamentos.objects.all()  # Recupera todos os equipamentos do banco de dados
-    return render(request, 'gerenciador_industriawayne/listar_equipamentos.html', {'equipamentos': equipamentos})
+    if not request.user.has_perm('gerenciador_industriawayne.view_equipamentos'):
+        # Adiciona uma mensagem de erro
+        messages.error(request, 'Você não tem permissão para acessar esta página.')
+        return redirect('home')
+    
+    else:
+        equipamentos = Equipamentos.objects.all()  # Recupera todos os equipamentos do banco de dados
+        return render(request, 'gerenciador_industriawayne/listar_equipamentos.html', {'equipamentos': equipamentos})
 
 class editar_equipamento(UpdateView):
     model = Equipamentos
@@ -38,34 +48,46 @@ class editar_equipamento(UpdateView):
         return super().form_valid(form)
     
 @login_required(login_url='/login/login')
-@permission_required('gerenciador_industriawayne.view_equipamentos', raise_exception=True)
 def remover_equipamento(request, equipamento_id):
-    equipamento = get_object_or_404(Equipamentos, id=equipamento_id)
+    if not request.user.has_perm('gerenciador_industriawayne.delete_equipamentos'):
+        # Adiciona uma mensagem de erro
+        messages.error(request, 'Você não tem permissão para acessar esta página.')
+        return redirect('home')
+    else:
+        equipamento = get_object_or_404(Equipamentos, id=equipamento_id)
     
-    if request.method == 'POST':
-        equipamento.delete()
-        return redirect('listar_equipamentos')  # Redireciona para a lista de equipamentos após exclusão
+        if request.method == 'POST':
+            equipamento.delete()
+            return redirect('listar_equipamentos')  # Redireciona para a lista de equipamentos após exclusão
     
-    return render(request, 'gerenciador_industriawayne/remover_equipamento.html', {'equipamento': equipamento})
+        return render(request, 'gerenciador_industriawayne/remover_equipamento.html', {'equipamento': equipamento})
 
 @login_required(login_url='/login/login')
-@permission_required('gerenciador_industriawayne.view_equipamentos', raise_exception=True)
 def cadastrar_inimigos(request):
-    form = InimigosForms(request.POST, request.FILES)
-    if form.is_valid():
-        form.save()
-        messages.success(request, 'Inimigo Cadastrado com sucesso!')
-    return render(request, 'gerenciador_industriawayne/cadastrar_inimigos.html', {'form': form})
+    if not request.user.has_perm('gerenciador_industriawayne.add_inimigos'):
+        # Adiciona uma mensagem de erro
+        messages.error(request, 'Você não tem permissão para acessar esta página.')
+        return redirect('home')
+    else:
+        form = InimigosForms(request.POST, request.FILES)
+        if form.is_valid():
+            form.save()
+            messages.success(request, 'Inimigo Cadastrado com sucesso!')
+        return render(request, 'gerenciador_industriawayne/cadastrar_inimigos.html', {'form': form})
 
 @login_required(login_url='/login/login')
 def listar_inimigos(request):
-    inimigos = Inimigos.objects.all()  # Recupera todos os equipamentos do banco de dados
-    paginator = Paginator(inimigos, 3)
-    page_number = request.GET.get('page')  # Captura o número da página na URL
-    page_obj = paginator.get_page(page_number)		
-    return render(request, 'gerenciador_industriawayne/listar_inimigos.html', {'inimigos': inimigos, 'page_obj': page_obj})
+    if not request.user.has_perm('gerenciador_industriawayne.view_inimigos'):
+        # Adiciona uma mensagem de erro
+        messages.error(request, 'Você não tem permissão para acessar esta página.')
+        return redirect('home')
+    else:
+        inimigos = Inimigos.objects.all()  # Recupera todos os equipamentos do banco de dados
+        paginator = Paginator(inimigos, 3)
+        page_number = request.GET.get('page')  # Captura o número da página na URL
+        page_obj = paginator.get_page(page_number)		
+        return render(request, 'gerenciador_industriawayne/listar_inimigos.html', {'inimigos': inimigos, 'page_obj': page_obj})
 
-@permission_required('gerenciador_industriawayne.view_equipamentos', raise_exception=True)
 class editar_inimigo(UpdateView):
     model = Inimigos
     fields = ['nome_inimigo', 'sexo', 'super_poder', 'armas', 'grau_de_perigo', 'descricao', 'capturado','data_captura','localizacao', 'imagem']
@@ -77,31 +99,41 @@ class editar_inimigo(UpdateView):
         return super().form_valid(form)
 
 @login_required(login_url='/login/login')
-@permission_required('gerenciador_industriawayne.view_equipamentos', raise_exception=True)
 def remover_inimigo(request, inimigo_id):
-    inimigo = get_object_or_404(Inimigos, id=inimigo_id)
-    
-    if request.method == 'POST':
-        inimigo.delete()
-        return redirect('listar_inimigos')  # Redireciona para a lista de inimigos após exclusão
-    
-    return render(request, 'gerenciador_industriawayne/remover_inimigo.html', {'inimigo': inimigo})
+    if not request.user.has_perm('gerenciador_industriawayne.delete_inimigos'):
+    # Adiciona uma mensagem de erro
+        messages.error(request, 'Você não tem permissão para acessar esta página.')
+        return redirect('home')
+    else:
+        inimigo = get_object_or_404(Inimigos, id=inimigo_id)
+        if request.method == 'POST':
+            inimigo.delete()
+            return redirect('listar_inimigos')  # Redireciona para a lista de inimigos após exclusão
+        return render(request, 'gerenciador_industriawayne/remover_inimigo.html', {'inimigo': inimigo})
 
 @login_required(login_url='/login/login')
-@permission_required('gerenciador_industriawayne.view_equipamentos', raise_exception=True)
 def cadastrar_meta(request):
-    form = MetasForms(request.POST, request.FILES)
-    if form.is_valid():
-        form.save()
-        messages.success(request, 'Meta Cadastrada com sucesso!')
-    return render(request, 'gerenciador_industriawayne/cadastrar_metas.html', {'form': form})
+    if not request.user.has_perm('gerenciador_industriawayne.add_metas'):
+        # Adiciona uma mensagem de erro
+        messages.error(request, 'Você não tem permissão para acessar esta página.')
+        return redirect('home')
+    else:
+        form = MetasForms(request.POST, request.FILES)
+        if form.is_valid():
+            form.save()
+            messages.success(request, 'Meta Cadastrada com sucesso!')
+        return render(request, 'gerenciador_industriawayne/cadastrar_metas.html', {'form': form})
 
 @login_required(login_url='/login/login')
 def listar_metas(request):
-    metas = Metas.objects.all()  # Recupera todos os equipamentos do banco de dados
-    return render(request, 'gerenciador_industriawayne/listar_metas.html', {'metas': metas})
+    if not request.user.has_perm('gerenciador_industriawayne.view_metas'):
+        # Adiciona uma mensagem de erro
+        messages.error(request, 'Você não tem permissão para acessar esta página.')
+        return redirect('home')
+    else:
+        metas = Metas.objects.all()  # Recupera todos os equipamentos do banco de dados
+        return render(request, 'gerenciador_industriawayne/listar_metas.html', {'metas': metas})
 
-@permission_required('gerenciador_industriawayne.view_equipamentos', raise_exception=True)
 class editar_meta(UpdateView):
     model = Metas
     fields = ['nome_meta', 'responsavel', 'descricao', 'data_prazo','status']
@@ -113,12 +145,17 @@ class editar_meta(UpdateView):
         return super().form_valid(form)
 
 @login_required(login_url='/login/login')
-@permission_required('gerenciador_industriawayne.view_equipamentos', raise_exception=True)
+
 def remover_meta(request, meta_id):
-    meta = get_object_or_404(Metas, id=meta_id)
-    
-    if request.method == 'POST':
-        meta.delete()
-        return redirect('listar_metas')  # Redireciona para a lista de metas após exclusão
-    
-    return render(request, 'gerenciador_industriawayne/remover_meta.html', {'meta': meta})
+    if not request.user.has_perm('gerenciador_industriawayne.delete_metas'):
+        # Adiciona uma mensagem de erro
+        messages.error(request, 'Você não tem permissão para acessar esta página.')
+        return redirect('home')
+    else:
+        meta = get_object_or_404(Metas, id=meta_id)
+        
+        if request.method == 'POST':
+            meta.delete()
+            return redirect('listar_metas')  # Redireciona para a lista de metas após exclusão
+        
+        return render(request, 'gerenciador_industriawayne/remover_meta.html', {'meta': meta})
